tools/generate-lesson/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "generate-lesson"
version = "0.1.0"
dependencies = ["pyyaml>=6.0.1"]

[project.scripts]
generate-lesson = "generate_lesson.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["generate_lesson*"]
